+++ src/main/java/racing/web/controller/dto/request/validator/RacingGameInfoRequestValidator.java
@@ -1,15 +1,15 @@
 package racing.web.controller.dto.request.validator;
 
-import static racing.common.exception.BusinessExceptionType.CAR_NAME_BLANK;
-import static racing.common.exception.BusinessExceptionType.CAR_NAME_DUPLICATION;
-import static racing.common.exception.BusinessExceptionType.CAR_NAME_INVALID_RANGE;
+import static racing.web.exception.RacingGameWebExceptionType.CAR_NAME_BLANK;
+import static racing.web.exception.RacingGameWebExceptionType.CAR_NAME_DUPLICATION;
+import static racing.web.exception.RacingGameWebExceptionType.CAR_NAME_INVALID_RANGE;
 
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.stream.Collectors;
-import racing.common.exception.BusinessException;
 import racing.web.controller.dto.request.RacingGameInfoRequest;
+import racing.web.exception.RacingGameWebException;
 
 public class RacingGameInfoRequestValidator implements RequestValidator<RacingGameInfoRequest> {
 
@@ -37,19 +37,19 @@ public class RacingGameInfoRequestValidator implements RequestValidator<RacingGa
         int distinctSize = new HashSet<>(names).size();
 
         if (names.size() != distinctSize) {
-            throw new BusinessException(CAR_NAME_DUPLICATION);
+            throw new RacingGameWebException(CAR_NAME_DUPLICATION);
         }
     }
 
     private void validateNotBlank(String name) {
         if (name == null || name.isBlank()) {
-            throw new BusinessException(CAR_NAME_BLANK);
+            throw new RacingGameWebException(CAR_NAME_BLANK);
         }
     }
 
     private void validateLengthInRange(String name) {
         if (name.length() < 1 || 5 < name.length()) {
-            throw new BusinessException(CAR_NAME_INVALID_RANGE);
+            throw new RacingGameWebException(CAR_NAME_INVALID_RANGE);
         }
     }
 }
